+++ src/app/features/posts/post-routing.module.ts
@@ -1,7 +1,8 @@
 import {NgModule} from '@angular/core';
 import {RouterModule, Routes} from '@angular/router';
 import {PostDetailsComponent} from './post-details/post-details.component';
-import {PostsComponent} from './posts.component';
+import {PostsComponent} from './posts-container/posts.component';
+
 
 
 const routes: Routes = [
