+++ src/app/features/posts/posts.module.ts
@@ -1,6 +1,6 @@
 import {NgModule} from '@angular/core';
 import {CommonModule} from '@angular/common';
-import {PostsComponent} from './posts.component';
+import {PostsComponent} from './posts-container/posts.component';
 import {UcfirstPipe} from 'src/app/core/pipes/ucfirst.pipe';
 import {BrowserAnimationsModule} from '@angular/platform-browser/animations';
 import {MatCardModule} from '@angular/material/card';
@@ -11,6 +11,7 @@ import {RouterModule} from '@angular/router';
 import {PostDetailsComponent} from 'src/app/features/posts/post-details/post-details.component';
 import {PostRoutingModule} from './post-routing.module';
 import {CommentsModule} from '../comments/comments.module';
+
 
 
 @NgModule({
